Add item order messages to elaboration context
To avoid the need to deal with them separately in the driver.
